don't mess with data types, just cast to int like a sane person

+++ prx/rc1/multiplayer/GameClient.cpp
@@ -253,10 +253,10 @@ void GameClient::update_set_state(MPPacketSetState* packet) {
         }
         case MP_STATE_TYPE_SET_BOLTS: {
             uint32_t value = (uint32_t)packet->value;
-            *(uint32_t*)player_bolts += ((value>>24)&0xff) | // move byte 3 to byte 0
-                                        ((value<<8)&0xff0000) | // move byte 1 to byte 2
-                                        ((value>>8)&0xff00) | // move byte 2 to byte 1
-                                        ((value<<24)&0xff000000); // byte 0 to byte 3
+            *(int*)player_bolts += (int)(((value>>24)&0xff) | // move byte 3 to byte 0
+                                         ((value<<8)&0xff0000) | // move byte 1 to byte 2
+                                         ((value>>8)&0xff00) | // move byte 2 to byte 1
+                                         ((value<<24)&0xff000000)); // byte 0 to byte 3
             break;
         }
         case MP_STATE_TYPE_GIVE_BOLTS: {
